Updated insertion optimizations.
This commit replaces code that consolidated the search algorithms.
The performance of the consolidated code was not quite as good as
the code in this commit. The difference was large enough to
measure and thus warranted this change since the original intent
behind adding the search algorithms was performance. Unfortunately
if a problem is found with the algorithms there are two different
places to update. The es_sink and es_queue classes now have their
own search algorithms.

The default search behavior was changed to SEARCH_BINARY. The enum
es_search_styles was changed to es_search_behaviors to mimic the
existing es_queue_early_behaviors enum. es_search_styles is defined
in es_common.h.

There was a problem with the previous commit where the es_queue
search behavior was not being updated from the property on the
GRC sink block (although the es_sink search behavior was updated).
This problem has been fixed.