Remove dependency on Control Port backend.
It is not necessary to check for a Control Port backend (ICE/THRIFT).

The ENABLE_GR_CTRLPORT option allows this module to opt out of
participating in the GNU Radio block Control Port interface
(setup_rpc method). Whether or not Control Port works is entirely
dependent on the GNU Radio build and whether or not it was compiled
with Control Port support (including checking for the particular
backend required/requested); but that has little to do with this
module (gr-eventstream).